Add createBookmark' that returns a full bookmark, and make signature for deleteBookmark more honest.
The remote call to create a bookmark does not actually return the full
bookmark, but only the id, title, and url fields.  createBoomark'
returns the full new bookmark by doing two calls - 1 to create the mark,
then another to immediately fetch it.

deleteBookmark always returned an empty string, so () is more accurate.